Tidying up Codebase and Preperations
I started work tidying up the API routes, seperating the HTTP routing logic from the data layer.

The next immediate goal is to implement more deeply integrated logging for all of the API, including returing refrences to the session and error when a user gets an error, so we can investigate the cause.

After that, implementing the permission system is the most important, with all entitlements belonging to roles which belong to groups containing users, or roles can belong directly to users. Using super basic priority accounting will be where we start.